backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi (>=0.120.0,<0.121.0)",
    "uvicorn[standard] (>=0.38.0,<0.39.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "alembic (>=1.17.0,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pydantic-settings (>=2.11.0,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.12.3,<3.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)"
]

[dependency-groups]
dev = [
    "black (>=25.9.0,<26.0.0)",
    "ruff (>=0.14.2,<0.15.0)",
    "mypy (>=1.18.2,<2.0.0)",
    "pytest (>=8.4.2,<9.0.0)",
    "poethepoet (>=0.37.0,<0.38.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.poe.tasks]
lint = "ruff check app tests"
format = "black app tests"
check = ["lint", "format --check"]
